+++ src/android/io/sqlc/SQLiteAndroidDatabase.java
@@ -8,12 +8,16 @@ package io.sqlc;
 
 import android.annotation.SuppressLint;
 
-import android.database.Cursor;
-import android.database.CursorWindow;
-import android.database.sqlite.SQLiteCursor;
-import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteException;
-import android.database.sqlite.SQLiteStatement;
+//import android.database.Cursor;
+//import android.database.CursorWindow;
+//import android.database.sqlite.SQLiteCursor;
+//import android.database.sqlite.SQLiteDatabase;
+//import android.database.sqlite.SQLiteException;
+//import android.database.sqlite.SQLiteStatement;
+
+// SQLCipher version of database classes:
+import net.sqlcipher.*;
+import net.sqlcipher.database.*;
 
 import android.util.Base64;
 import android.util.Log;
@@ -24,7 +28,11 @@ import java.lang.Number;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.apache.cordova.CallbackContext;
+//import org.apache.cordova.CallbackContext;
+
+// NOTE: more than CordovaPlugin & CallbackContext needed to support
+// ...
+import org.apache.cordova.*;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -56,13 +64,26 @@ class SQLiteAndroidDatabase
      */
 
     /**
+     * to load native lib(s)
+     */
+    //@Override
+    static
+    public void initialize(CordovaInterface cordova) {
+        //super.initialize(cordova, webView);
+        //SQLiteDatabase.loadLibs(this.cordova.getActivity());
+        SQLiteDatabase.loadLibs(cordova.getActivity());
+    }
+
+    /**
+     *
      * Open a database.
      *
      * @param dbfile   The database File specification
      */
-    void open(File dbfile) throws Exception {
-        dbFile = dbfile; // for possible bug workaround
-        mydb = SQLiteDatabase.openOrCreateDatabase(dbfile, null);
+    void open(File dbfile, String key) throws Exception {
+        dbFile = dbfile; // for possible bug workaround (NOT NEEDED in this version)
+        //mydb = SQLiteDatabase.openOrCreateDatabase(dbfile, null);
+        mydb = SQLiteDatabase.openOrCreateDatabase(dbfile, key, null);
     }
 
     /**
@@ -75,10 +96,12 @@ class SQLiteAndroidDatabase
         }
     }
 
+    /* NOT NEEDED in this version:
     void bugWorkaround() throws Exception {
         this.closeDatabaseNow();
         this.open(dbFile);
     }
+    */
 
     /**
      * Executes a batch request and sends the results via cbc.
@@ -105,8 +128,7 @@ class SQLiteAndroidDatabase
         JSONArray batchResults = new JSONArray();
 
         for (int i = 0; i < len; i++) {
-            int rowsAffectedCompat = 0;
-            boolean needRowsAffectedCompat = false;
+            // NOTE: no need for rowsAffectedCompat hack with SQLCipher for Android
             query_id = queryIDs[i];
 
             JSONObject queryResult = null;
@@ -120,38 +142,32 @@ class SQLiteAndroidDatabase
                 QueryType queryType = getQueryType(query);
 
                 if (queryType == QueryType.update || queryType == queryType.delete) {
-                    if (android.os.Build.VERSION.SDK_INT >= 11) {
-                        SQLiteStatement myStatement = mydb.compileStatement(query);
+                    // NOTE: SQLCipher for Android provides consistent SQLiteStatement.executeUpdateDelete();
+                    // no need for rowsAffectedCompat hack.
+                    SQLiteStatement myStatement = mydb.compileStatement(query);
 
-                        if (jsonparams != null) {
-                            bindArgsToStatement(myStatement, jsonparams[i]);
-                        }
+                    if (jsonparams != null) {
+                        bindArgsToStatement(myStatement, jsonparams[i]);
+                    }
 
-                        int rowsAffected = -1; // (assuming invalid)
+                    long rowsAffected = -1; // (assuming invalid)
 
-                        // Use try & catch just in case android.os.Build.VERSION.SDK_INT >= 11 is lying:
-                        try {
-                            rowsAffected = myStatement.executeUpdateDelete();
-                            // Indicate valid results:
-                            needRawQuery = false;
-                        } catch (SQLiteException ex) {
-                            // Indicate problem & stop this query:
-                            ex.printStackTrace();
-                            errorMessage = ex.getMessage();
-                            Log.v("executeSqlBatch", "SQLiteStatement.executeUpdateDelete(): Error=" + errorMessage);
-                            needRawQuery = false;
-                        } catch (Exception ex) {
-                            // Assuming SDK_INT was lying & method not found:
-                            // do nothing here & try again with raw query.
-                        }
+                    try {
+                        rowsAffected = myStatement.executeUpdateDelete();
+                        // Indicate valid results:
+                        needRawQuery = false;
+                    } catch (SQLiteException ex) {
+                        // Indicate problem & stop this query:
+                        ex.printStackTrace();
+                        errorMessage = ex.getMessage();
+                        Log.v("executeSqlBatch", "SQLiteStatement.executeUpdateDelete(): Error=" + errorMessage);
+                        // stop the query in case of error:
+                        needRawQuery = false;
+                    }
 
-                        if (rowsAffected != -1) {
-                            queryResult = new JSONObject();
-                            queryResult.put("rowsAffected", rowsAffected);
-                        }
-                    } else { // pre-honeycomb
-                        rowsAffectedCompat = countRowsAffectedCompat(queryType, query, jsonparams, mydb, i);
-                        needRowsAffectedCompat = true;
+                    if (rowsAffected != -1) {
+                        queryResult = new JSONObject();
+                        queryResult.put("rowsAffected", rowsAffected);
                     }
                 }
 
@@ -231,10 +247,6 @@ class SQLiteAndroidDatabase
                 // raw query for other statements:
                 if (needRawQuery) {
                     queryResult = this.executeSqlStatementQuery(mydb, query, jsonparams[i], cbc);
-
-                    if (needRowsAffectedCompat) {
-                        queryResult.put("rowsAffected", rowsAffectedCompat);
-                    }
                 }
             } catch (Exception ex) {
                 ex.printStackTrace();
@@ -270,81 +282,6 @@ class SQLiteAndroidDatabase
         }
 
         cbc.success(batchResults);
-    }
-
-    private int countRowsAffectedCompat(QueryType queryType, String query, JSONArray[] jsonparams,
-                                         SQLiteDatabase mydb, int i) throws JSONException {
-        // quick and dirty way to calculate the rowsAffected in pre-Honeycomb.  just do a SELECT
-        // beforehand using the same WHERE clause. might not be perfect, but it's better than nothing
-        Matcher whereMatcher = WHERE_CLAUSE.matcher(query);
-
-        String where = "";
-
-        int pos = 0;
-        while (whereMatcher.find(pos)) {
-            where = " WHERE " + whereMatcher.group(1);
-            pos = whereMatcher.start(1);
-        }
-        // WHERE clause may be omitted, and also be sure to find the last one,
-        // e.g. for cases where there's a subquery
-
-        // bindings may be in the update clause, so only take the last n
-        int numQuestionMarks = 0;
-        for (int j = 0; j < where.length(); j++) {
-            if (where.charAt(j) == '?') {
-                numQuestionMarks++;
-            }
-        }
-
-        JSONArray subParams = null;
-
-        if (jsonparams != null) {
-            // only take the last n of every array of sqlArgs
-            JSONArray origArray = jsonparams[i];
-            subParams = new JSONArray();
-            int startPos = origArray.length() - numQuestionMarks;
-            for (int j = startPos; j < origArray.length(); j++) {
-                subParams.put(j - startPos, origArray.get(j));
-            }
-        }
-
-        if (queryType == QueryType.update) {
-            Matcher tableMatcher = UPDATE_TABLE_NAME.matcher(query);
-            if (tableMatcher.find()) {
-                String table = tableMatcher.group(1);
-                try {
-                    SQLiteStatement statement = mydb.compileStatement(
-                            "SELECT count(*) FROM " + table + where);
-
-                    if (subParams != null) {
-                        bindArgsToStatement(statement, subParams);
-                    }
-
-                    return (int)statement.simpleQueryForLong();
-                } catch (Exception e) {
-                    // assume we couldn't count for whatever reason, keep going
-                    Log.e(SQLiteAndroidDatabase.class.getSimpleName(), "uncaught", e);
-                }
-            }
-        } else { // delete
-            Matcher tableMatcher = DELETE_TABLE_NAME.matcher(query);
-            if (tableMatcher.find()) {
-                String table = tableMatcher.group(1);
-                try {
-                    SQLiteStatement statement = mydb.compileStatement(
-                            "SELECT count(*) FROM " + table + where);
-                    bindArgsToStatement(statement, subParams);
-
-                    return (int)statement.simpleQueryForLong();
-                } catch (Exception e) {
-                    // assume we couldn't count for whatever reason, keep going
-                    Log.e(SQLiteAndroidDatabase.class.getSimpleName(), "uncaught", e);
-
-                }
-            }
-        }
-
-        return 0;
     }
 
     private void bindArgsToStatement(SQLiteStatement myStatement, JSONArray sqlArgs) throws JSONException {
@@ -406,17 +343,8 @@ class SQLiteAndroidDatabase
                     for (int i = 0; i < colCount; ++i) {
                         key = cur.getColumnName(i);
 
-                        if (android.os.Build.VERSION.SDK_INT >= 11) {
-
-                            // Use try & catch just in case android.os.Build.VERSION.SDK_INT >= 11 is lying:
-                            try {
-                                bindPostHoneycomb(row, key, cur, i);
-                            } catch (Exception ex) {
-                                bindPreHoneycomb(row, key, cur, i);
-                            }
-                        } else {
-                            bindPreHoneycomb(row, key, cur, i);
-                        }
+                        // Always valid for SQLCipher for Android:
+                        bindPostHoneycomb(row, key, cur, i);
                     }
 
                     rowsArrayResult.put(row);
@@ -440,7 +368,10 @@ class SQLiteAndroidDatabase
         return rowsResult;
     }
 
-    @SuppressLint("NewApi")
+    /**
+     * bindPostHoneycomb - always valid for SQLCipher for Android
+     *
+     */
     private void bindPostHoneycomb(JSONObject row, String key, Cursor cur, int i) throws JSONException {
         int curType = cur.getType(i);
 
@@ -461,26 +392,6 @@ class SQLiteAndroidDatabase
             default: /* (not expected) */
                 row.put(key, cur.getString(i));
                 break;
-        }
-    }
-
-    private void bindPreHoneycomb(JSONObject row, String key, Cursor cursor, int i) throws JSONException {
-        // Since cursor.getType() is not available pre-honeycomb, this is
-        // a workaround so we don't have to bind everything as a string
-        // Details here: http://stackoverflow.com/q/11658239
-        SQLiteCursor sqLiteCursor = (SQLiteCursor) cursor;
-        CursorWindow cursorWindow = sqLiteCursor.getWindow();
-        int pos = cursor.getPosition();
-        if (cursorWindow.isNull(pos, i)) {
-            row.put(key, JSONObject.NULL);
-        } else if (cursorWindow.isLong(pos, i)) {
-            row.put(key, cursor.getLong(i));
-        } else if (cursorWindow.isFloat(pos, i)) {
-            row.put(key, cursor.getDouble(i));
-        } else if (cursorWindow.isBlob(pos, i)) {
-            row.put(key, new String(Base64.encode(cursor.getBlob(i), Base64.DEFAULT)));
-        } else { // string
-            row.put(key, cursor.getString(i));
         }
     }