+++ CHANGES.md
@@ -4,9 +4,15 @@
 
 XXX TBD
 
+### cordova-sqlite-storage 1.2.1
+
+- Close Android SQLiteStatement after INSERT/UPDATE/DELETE
+- Specify minimum Cordova version 6.0.0
+- Lawnchair adapter fix: Changed remove method to work with key array
+
 ### x.x.x-common-dev
 
-- Support Windows 8.1/Windows Phone 8.1 along with Windows 10
+- TODO (disabled): Support Windows 8.1/Windows Phone 8.1 along with Windows 10
 - Fix PCH issue with Debug Win32 UWP (Windows 10) build
 
 ### cordova-sqlite-storage 1.2.0